+++ src/test/kotlin/repositories/RepositoryPublisherFactoryTest.kt
@@ -6,20 +6,20 @@ import dev.zuccaops.helpers.VersionResolver
 import dev.zuccaops.repositories.RepositoryConstants
 import dev.zuccaops.repositories.RepositoryPublisherFactory
 import dev.zuccaops.repositories.central.MavenCentralRepositoryPublisher
-import dev.zuccaops.repositories.central.SonatypeRepositoryPublisher
+import dev.zuccaops.repositories.central.NexusRepositoryPublisher
 import dev.zuccaops.repositories.local.LocalRepositoryPublisher
 import dev.zuccaops.repositories.remote.RemoteRepositoryPublisher
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.mockkConstructor
-import org.gradle.api.Project
 import org.gradle.internal.impldep.org.junit.Assert.assertTrue
 import org.junit.jupiter.api.Test
+import testutil.TestProjectFactory
 
 class RepositoryPublisherFactoryTest {
 
 
-    private val project = mockk<Project>(relaxed = true)
+    private val project = TestProjectFactory.create()
 
     @Test
     fun `should return RemoteRepositoryPublisher in dev mode`() {
@@ -97,7 +97,7 @@ class RepositoryPublisherFactoryTest {
         val publisher = RepositoryPublisherFactory.get(project, config)
 
         // then
-        assertTrue(publisher is SonatypeRepositoryPublisher)
+        assertTrue(publisher is NexusRepositoryPublisher)
     }