+++ src/test/kotlin/repositories/central/NexusRepositoryPublisherTest.kt
@@ -1,36 +1,36 @@
 package repositories.central
 
 import dev.zuccaops.helpers.VersionResolver
-import dev.zuccaops.repositories.central.SonatypeRepositoryPublisher
+import dev.zuccaops.repositories.central.NexusRepositoryPublisher
 import io.mockk.every
 import io.mockk.mockk
-import org.gradle.api.Project
 import org.gradle.api.artifacts.ArtifactRepositoryContainer.MAVEN_CENTRAL_URL
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
+import testutil.TestProjectFactory
 
-class SonatypeRepositoryPublisherTest {
+class NexusRepositoryPublisherTest {
 
     @Test
     fun `shouldSign always returns true`() {
-        val publisher = SonatypeRepositoryPublisher(mockk(), mockk(), "customCommand")
+        val publisher = NexusRepositoryPublisher(TestProjectFactory.create(), mockk(), "customCommand")
         assertTrue(publisher.shouldSign())
     }
 
     @Test
     fun `getUri should return correct artifact path`() {
         // given
-        val mockProject = mockk<Project> {
-            every { group } returns "com.example.library"
-            every { name } returns "my-lib"
-        }
+        val mockProject = TestProjectFactory.create()
+
+        every { mockProject.group } returns "com.example.library"
+        every { mockProject.name } returns "my-lib"
 
         val mockResolver = mockk<VersionResolver> {
             every { getVersion() } returns "2.3.4"
         }
 
-        val publisher = SonatypeRepositoryPublisher(mockProject, mockResolver, "someTask")
+        val publisher = NexusRepositoryPublisher(mockProject, mockResolver, "someTask")
 
         // when
         val uri = publisher.getUri()